requirements.txt
Flask==2.0.1
pandas==1.3.3
comtypes==1.1.14
python-pptx==0.6.21
Werkzeug==2.0.1
Pillow==10.1.0
python-dotenv==1.0.0 